+ Refactor some class locations from the main UI package into the UI Utils package
