+++ MekHQ/src/mekhq/campaign/Campaign.java
@@ -160,7 +160,7 @@ import mekhq.campaign.work.IAcquisitionWork;
 import mekhq.campaign.work.IMedicalWork;
 import mekhq.campaign.work.IPartWork;
 import mekhq.campaign.work.Modes;
-import mekhq.gui.PortraitFileFactory;
+import mekhq.gui.utilities.PortraitFileFactory;
 
 import org.w3c.dom.DOMException;
 import org.w3c.dom.Document;
