+++ MekHQ/src/mekhq/gui/adapter/ServicedUnitsTableMouseAdapter.java
@@ -17,7 +17,7 @@ import mekhq.campaign.parts.Part;
 import mekhq.campaign.parts.equipment.AmmoBin;
 import mekhq.campaign.unit.Unit;
 import mekhq.gui.CampaignGUI;
-import mekhq.gui.MenuScroller;
+import mekhq.gui.utilities.MenuScroller;
 import mekhq.gui.utilities.StaticChecks;
 
 public class ServicedUnitsTableMouseAdapter extends MouseInputAdapter
